requirements.txt
flask==2.0.1
requests==2.26.0
flask-cors==3.0.10
gunicorn==20.1.0
python-telegram-bot==20.6